removed redundant code

+++ include/cpga/utilities/csv_reader.hpp
@@ -14,7 +14,7 @@ namespace cpga {
 namespace utilities {
 class csv_reader {
  public:
-  static auto read_double(const std::string &file, size_t rows, size_t cols, char delim = ',') {
+  static auto read_double(const std::string &file, int rows, int cols, char delim = ',') {
     std::vector<std::vector<double>> data(rows, std::vector<double>(cols));
     std::ifstream ifs(file);
 
@@ -23,13 +23,13 @@ class csv_reader {
     }
 
     std::string s;
-    for (size_t i = 0; i < rows; ++i) {
+    for (int i = 0; i < rows; ++i) {
       if (!std::getline(ifs, s)) {
         throw std::runtime_error("No more rows to read");
       }
       std::istringstream iss(s);
       std::string token;
-      for (size_t j = 0; j < cols; ++j) {
+      for (int j = 0; j < cols; ++j) {
         if (!std::getline(iss, token, delim)) {
           throw std::runtime_error("No more columns to read");
         }
